requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
anyjson==0.3.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
bcrypt==3.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.6
defusedxml==0.7.1
distlib==0.3.1
Django==3.1
django-admin==2.0.1
django-allauth==0.44.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-formtools==2.3
django-phonenumber-field==5.1.0
django-picklefield==0.2.1
django-registration==2.0.1
django-six==1.0.4
django-tinymce==3.2.0
djangorestframework==3.12.2
filelock==3.0.12
future==0.18.2
gevent==21.1.2
greenlet==1.0.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==3.7.3
importlib-resources==5.1.2
ipaddr==2.2.0
jsonpickle==2.0.0
kombu==5.1.0
lxml==4.2.1
multidict==5.1.0
ncclient==0.6.10
nxpy==0.4.0
oauthlib==3.1.0
paramiko==2.7.2
phonenumbers==8.12.22
ply==3.11
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
pycryptodomex==3.10.1
PyJWT==2.1.0
PyNaCl==1.4.0
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.1
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.1
qrcode==6.1
requests==2.25.1
requests-oauthlib==1.3.0
screen==1.0.1
six==1.15.0
slackclient==2.9.3
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.3
wcwidth==0.2.5
xlwt==1.3.0
yarl==1.6.3
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.3.0
redis==3.5.3
django-registration-redux==2.9
pybix==0.0.8
pyzabbix==1.0.0
netaddr==0.8.0
pillow==8.3.1
django-dbbackup==3.3.0
dj-webhooks==0.2.1
django-braces==1.14.0
django-sass==1.1.0
django-simple-history==3.0.0
